Add comprehensive test suite for SecretsStore functionality
- Implement tests for API adapter in api_adapter_test.go, validating SecretReader, SecretWriter, Authenticator, UserManager, and FullService interfaces.
- Create tests for key management in crypto_keys_test.go, ensuring key creation, loading, and error handling for invalid keys.
- Add rotation and cleanup tests in rotation_cleanup_test.go to verify backup management and restoration processes.
- Introduce atomic operation tests in secrets_atomic_test.go to ensure atomicity and cleanup of temporary files during key rotations.
- Implement backup and restore tests in secrets_backup_test.go to validate backup creation on overwrite and deletion.
- Add error handling tests in secrets_error_test.go for corrupted and missing JSON files.
- Create filesystem tests in secrets_filesystem_test.go to ensure proper file permissions, atomicity, and secret management across restarts.
