+++ internal/api_adapter_test.go
